feat: Add a service module and make the service generic over the protocol. other updates in benches
