+++ scheduler/src/solvers/mod.rs
@@ -1,20 +1,38 @@
 mod greedy;
 use crate::config::Settings;
-use crate::solver::Solver;
+use crate::Error;
+use crate::{DeviceId, Pid, ResourceAlloc, Resources, TaskRequirements, TaskState};
 pub use greedy::GreedySolver;
+use std::collections::{HashMap, VecDeque};
 
 pub(crate) fn create_solver(_config: Option<&Settings>) -> Box<dyn Solver> {
     Box::new(GreedySolver)
+}
+// Trait that is implemented by any object that can be used as a solver
+pub trait Solver {
+    fn solve_job_schedule(
+        &mut self,
+        current_state: &HashMap<Pid, TaskState>,
+        scheduler_settings: &Settings,
+    ) -> Result<VecDeque<Pid>, Error>;
+
+    fn allocate_task(
+        &mut self,
+        resources: &Resources,
+        requirements: &TaskRequirements,
+        restrictions: &Option<Vec<DeviceId>>,
+        task_state: &HashMap<Pid, TaskState>,
+    ) -> Option<ResourceAlloc>;
 }
 
 #[cfg(dummy_devices)]
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::solver::TaskState;
+    use crate::TaskState;
     use crate::{
-        list_devices, solver::Resources, DeviceId, Pid, ResourceAlloc, ResourceMemory, ResourceReq,
-        ResourceState, ResourceType, TaskRequirements,
+        list_devices, DeviceId, Pid, ResourceAlloc, ResourceMemory, ResourceReq, ResourceState,
+        ResourceType, Resources, TaskRequirements,
     };
     use chrono::Utc;
     use std::collections::HashMap;
